bugfix(radar): Fixes expanding the VC pool
