+++ tests/dummy/app/routes/examples/infinite-scroll/template.hbs
@@ -1,44 +1,41 @@
 <div class="container-fluid">
-    <div class="row">
-        <div class="col-sm-5">
-            <h3>Demo</h3>
-          {{!- BEGIN-SNIPPET infinite-scroll-example }}
-            <div class="table-wrapper dark">
-              {{#vertical-collection model.numbers
-                minHeight=someProperty
-                alwaysRemeasure=true
-                debug=true
-                firstReached="loadAbove"
-                lastReached="loadBelow"
-                as |item index heightDidChange|
-              }}
-                {{number-slide
-                  item=item
-                  index=index
-                  didResize=heightDidChange
-                }}
-              {{/vertical-collection}}
-            </div>
-          {{!- END-SNIPPET }}
+  <div class="row">
+    <div class="col-sm-5">
+      <h3>Demo</h3>
+      {{!- BEGIN-SNIPPET infinite-scroll-example }}
+        <div class="table-wrapper dark">
+          {{#vertical-collection model.numbers
+            minHeight=40
+            alwaysRemeasure=true
+            firstReached="loadAbove"
+            lastReached="loadBelow"
+            as |item index heightDidChange|
+          }}
+            {{number-slide
+              item=item
+              index=index
+              isDynamic=false
+            }}
+          {{/vertical-collection}}
         </div>
-        <div class="col-sm-7">
-            <h3>Infinite Scroll</h3>
-            <p>
-                The Vertical Collection can be used to quickly build a robust infinite scroll.
-            </p>
-          <p>
-              Your infinite scroll can be bi-directional, loading new content above or
-              below.
-          </p>
-          <p>
-              Combined with `idForFirstItem`, this makes it very easy to load a user
-              in-media-res: ideal for caching their position in a long feed for when
-              they return.
-          </p>
-            <h4>Code for Demo</h4>
-
-          <button {{action "setMinHeight"}}>Do it </button>
-          {{!-- {{code-snippet name="infinite-scroll-example.hbs"}} --}}
-        </div>
+      {{!- END-SNIPPET }}
     </div>
+    <div class="col-sm-7">
+      <h3>Infinite Scroll</h3>
+      <p>
+        The Vertical Collection can be used to quickly build a robust infinite scroll.
+      </p>
+      <p>
+        Your infinite scroll can be bi-directional, loading new content above or
+        below.
+      </p>
+      <p>
+        Combined with `idForFirstItem`, this makes it very easy to load a user
+        in-media-res: ideal for caching their position in a long feed for when
+        they return.
+      </p>
+      <h4>Code for Demo</h4>
+      {{!-- {{code-snippet name="infinite-scroll-example.hbs"}} --}}
+    </div>
+  </div>
 </div>
