+++ tests/dummy/app/styles/app.css
@@ -88,7 +88,7 @@ number-slide {
   width: 250px;
   height: 4em;
   line-height: 4em;
-  margin: 10px auto 10px;
+  margin: 10px auto;
   background: #1f2f30;
   border: 1px solid #bbb;
   box-shadow: 1px 1px 5px 0 #d0d0d0;
